+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/AbstractIrDumpTest.kt
@@ -2,8 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.metro.compiler
 
+import kotlin.getValue
 import org.jetbrains.kotlin.config.JvmTarget
 import org.jetbrains.kotlin.test.builders.TestConfigurationBuilder
+import org.jetbrains.kotlin.test.directives.CodegenTestDirectives
 import org.jetbrains.kotlin.test.directives.CodegenTestDirectives.DUMP_IR
 import org.jetbrains.kotlin.test.directives.CodegenTestDirectives.DUMP_KT_IR
 import org.jetbrains.kotlin.test.directives.CodegenTestDirectives.IGNORE_DEXING
@@ -11,6 +13,7 @@ import org.jetbrains.kotlin.test.directives.ConfigurationDirectives.WITH_STDLIB
 import org.jetbrains.kotlin.test.directives.FirDiagnosticsDirectives.DISABLE_GENERATED_FIR_TAGS
 import org.jetbrains.kotlin.test.directives.JvmEnvironmentConfigurationDirectives.FULL_JDK
 import org.jetbrains.kotlin.test.directives.JvmEnvironmentConfigurationDirectives.JVM_TARGET
+import org.jetbrains.kotlin.test.directives.model.SimpleDirective
 import org.jetbrains.kotlin.test.runners.ir.AbstractFirLightTreeJvmIrTextTest
 import org.jetbrains.kotlin.test.services.KotlinStandardLibrariesPathProvider
 
@@ -35,9 +38,21 @@ open class AbstractIrDumpTest : AbstractFirLightTreeJvmIrTextTest() {
 
         -DUMP_IR
         +DUMP_KT_IR
+
+        // Disable the new SKIP_NEW_KOTLIN_REFLECT_COMPATIBILITY_CHECK, we don't need this here
+        // However, this fails in our infra _before_ we
+        SKIP_NEW_KOTLIN_REFLECT_COMPATIBILITY_CHECK_DIRECTIVE?.let { -it }
       }
 
       useMetaTestConfigurators(::MetroTestConfigurator)
     }
   }
+}
+
+private val SKIP_NEW_KOTLIN_REFLECT_COMPATIBILITY_CHECK_DIRECTIVE: SimpleDirective? by lazy {
+  CodegenTestDirectives::class
+    .java
+    .declaredMethods
+    .find { it.name == "getSKIP_NEW_KOTLIN_REFLECT_COMPATIBILITY_CHECK" }
+    ?.let { it.invoke(CodegenTestDirectives) as SimpleDirective }
 }